Decouple USB writes from memory writes (#11)
Ticket: #9

USB writes are now handled on a separate thread. The original write thread will submit a write task to the USB write thread if the USB write thread is available.

USB write tasks are not buffered. If the USB thread is not available, the image write will be skipped.